+++ Projectiles/Minions/CreeperProj.cs
@@ -36,8 +36,8 @@ namespace FargowiltasSouls.Projectiles.Minions
 			{
 				Main.dust[dust].velocity += projectile.velocity * 0.2f;
 			}
-			Main.dust[dust].position.X = projectile.Center.X + 4f + (float)Main.rand.Next(-2, 3);
-			Main.dust[dust].position.Y = projectile.Center.Y + (float)Main.rand.Next(-2, 3);
+			Main.dust[dust].position.X = projectile.Center.X + 4f + Main.rand.Next(-2, 3);
+			Main.dust[dust].position.Y = projectile.Center.Y + Main.rand.Next(-2, 3);
 			Main.dust[dust].noGravity = true;
 		}
