+++ Projectiles/SpearProjectileName.cs
@@ -1,8 +1,5 @@
 using System;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 using Terraria;
-using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace FargowiltasSouls.Projectiles
@@ -33,8 +30,8 @@ namespace FargowiltasSouls.Projectiles
             Main.player[projectile.owner].direction = projectile.direction;
             Main.player[projectile.owner].heldProj = projectile.whoAmI;
             Main.player[projectile.owner].itemTime = Main.player[projectile.owner].itemAnimation;
-            projectile.position.X = Main.player[projectile.owner].position.X + (float)(Main.player[projectile.owner].width / 2) - (float)(projectile.width / 2);
-            projectile.position.Y = Main.player[projectile.owner].position.Y + (float)(Main.player[projectile.owner].height / 2) - (float)(projectile.height / 2);
+            projectile.position.X = Main.player[projectile.owner].position.X + Main.player[projectile.owner].width / 2 - projectile.width / 2;
+            projectile.position.Y = Main.player[projectile.owner].position.Y + Main.player[projectile.owner].height / 2 - projectile.height / 2;
             projectile.position += projectile.velocity * projectile.ai[0]; if (projectile.ai[0] == 0f)
             {
                 projectile.ai[0] = 3f;
@@ -54,7 +51,7 @@ namespace FargowiltasSouls.Projectiles
                 projectile.Kill();
             }
 
-            projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 2.355f;
+            projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 2.355f;
             if (projectile.spriteDirection == -1)
             {
                 projectile.rotation -= 1.57f;
